hack: msm_zsl: always load phone firmware
As reported on xda, in the (apparently rare) event the sensor firmware
is newer than the phone firmware, the driver attempts to load the
firmware from /data. How the newer firmware gets there in the first
place is currently a mystery beknownst only to TouchWiz.

The user who made the report reported no ill effects from copying the
older firmware from /system/cameradata and moving it to /data

Always load the phone firmware then. But don't bypass the date checks
because they are also used to populate parts of the camera object and
also to load newer firmware manually from /data if it exists.

Change-Id: Icb54b762f827f5911c3078d5b22ff144c62259a1